\name{Stock-package}
\alias{Stock-package}
\alias{Stock}
\docType{package}
\title{
\packageTitle{Stock}
}
\description{
\packageDescription{Stock}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Stock}
\packageIndices{Stock}
An overview of how to use the package, including the most important functions
}
\author{Bianchi Federico, Corti Matteo, Giay Simone, Passano Andreas
\packageAuthor{Stock}

Maintainer: andreaspassano@gmail.com
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
